secure_voting_app-main/requirements.txt
streamlit
streamlit-oauth
requests
python-dotenv
